// src/models/__Inventory__.ts
export class __Inventory__ {
  inventoryId: string;
  productId: string;
  unit: string;
  description: string;
  unitPrice: number;
  quantityPurchased: number;
  quantityInStock: number;
  reorderLevel: number;
  reorderQuantity: number;
  reorderDate: Date;
  createdAt: Date;
  updatedAt: Date;
}